+++ plugin/controllers/ajax.py
@@ -20,6 +20,7 @@ from base import BaseController
 from time import mktime, localtime
 from models.locations import getLocations
 from twisted.web.resource import Resource
+import os
 
 try:
 	from boxbranding import getBoxType, getMachineName, getMachineBrand, getMachineBuild
@@ -187,9 +188,6 @@ class AjaxController(BaseController):
 		movies['sort'] = sorttype
 		return movies
 
-	def P_workinprogress(self, request):
-		return {}
-
 	def P_radio(self, request):
 		return {}
 
@@ -214,7 +212,10 @@ class AjaxController(BaseController):
 		}
 		ret['configsections'] = getConfigsSections()['sections']
 		if config.OpenWebif.webcache.theme.value:
-			ret['themes'] = config.OpenWebif.webcache.theme.choices
+			if os.path.exists(getPublicPath('themes')):
+				ret['themes'] = config.OpenWebif.webcache.theme.choices
+			else:
+				ret['themes'] = ['original','clear']
 			ret['theme'] = config.OpenWebif.webcache.theme.value
 		else:
 			ret['themes'] = []
@@ -293,7 +294,7 @@ class AjaxController(BaseController):
 		else:
 			auth=''
 		vxgenabled = False
-		if fileExists(getPublicPath("/js/media_player.pexe")):
+		if fileExists(getPublicPath("/vxg/media_player.pexe")):
 			vxgenabled = True
 		transcoding = getTranscodingSupport()
 		transcoder_port = 0