+++ CHANGES.md
@@ -1,5 +1,15 @@
-## Version 1.0.4
+## Version 1.1.0
 ## (in development)
+* reduce package size
+* allow remove mobile, webtv, themes without any impact to reduce the package size again
+-> remove the folder /public/webtv and/or mobile and/or themes 
+* add fancontrol2 webif to extras / #453
+* improve mutliepg / #462
+* improve channel list / #461
+* now button in timeline multiepg / #466
+* show timer in timeline multiepg / #460
+
+## Version 1.0.4
 * improve access security
 * add package manager
 * use jquery UI offline